# gateway/openapi.yaml
openapi: 3.1.0

info:
  title: MicroAI Paygate API
  version: "1.0.0"
  description: API documentation for MicroAI Paygate

paths:
  /healthz:
    get:
      summary: Health check
      description: Returns gateway health status
      responses:
        "200":
          description: Gateway is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  service:
                    type: string
                    example: gateway
                  status:
                    type: string
                    example: ok

  /api/ai/summarize:
    post:
      summary: Summarize text
      description: Proxies a text summarization request and enforces x402 payment
      parameters:
        - name: X-402-Signature
          in: header
          required: false
          description: EIP-712 signature authorizing payment
          schema:
            type: string
        - name: X-402-Nonce
          in: header
          required: false
          description: Nonce from 402 response
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  example: "Artificial intelligence is transforming software development."
      responses:
        "200":
          description: Summary generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: string
                    example: "AI is changing how software is built."
        "402":
          description: Payment required
        "403":
          description: Invalid signature
        "500":
          description: Server error

  /verify:
    post:
      summary: Verify signature (internal)
      description: Internal endpoint for EIP-712 verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                context:
                  type: object
                signature:
                  type: string
      responses:
        "200":
          description: Valid signature
        "403":
          description: Invalid signature